refactor : Test 내 Domain 파일 폴더별 분리
가독성 위해 , Test/Domain 내 파일들 폴더로 이동
