+++ src/test/java/christmas/domain/date/DateTest.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.date;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
