+++ src/test/java/christmas/domain/event/InProgressEventListTest.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.event;
 
 import christmas.domain.event.InProgressEventList;
 import christmas.factory.EventFactory;
