+++ src/test/java/christmas/domain/order/BillTest.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 import java.util.List;
