+++ src/test/java/christmas/domain/order/OrderTest.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
