+++ src/test/java/christmas/domain/reward/RewardTest.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.reward;
 
 import java.util.ArrayList;
 import java.util.List;
